requirements.txt
tensorflow-gpu
tensorflow-addons
tensorflow-hub
tensorflow-datasets
pydub
numpy
tqdm
larq
larq_zoo

